pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rvc"
version = "0.1.0"
dependencies = [
    "torch",
    "numpy",
    "ffmpeg-python",
    "scipy",
    "pydub",
    "torchaudio",
    "typer-slim",
    "tqdm",
]
authors = [
    { name = "Itzik", email = "74200683+itzik1058@users.noreply.github.com" },
]
license = { file = "LICENSE" }
readme = "README.md"

[project.optional-dependencies]
dev = ["mypy"]

[project.scripts]
rvc = "rvc.__main__:app"
convert-fairseq-wav2vec = "scripts.convert_fairseq_wav2vec:main"
export-rvc-project = "scripts.export_rvc_project:main"

[tool.mypy]
plugins = ["numpy.typing.mypy_plugin"]

disallow_any_generics = true
disallow_untyped_calls = true
disallow_untyped_defs = true
check_untyped_defs = true

[[tool.mypy.overrides]]
module = ["ffmpeg", "pydub.*", "scipy.*", "torchaudio.*", "fairseq.*"]
ignore_missing_imports = true
